+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.4.0] Reorienting Operations and Faces
+### Added
+- `Operation`: `get_face_near()` returns a Face whose center is nearest to a given point
+- `Face`:
+  - `shift()` method to rotate points around
+  - `reorient()` method that rotates points so that they start nearest to given position
+
 # [1.3.1] Projection Behaviour
 ### Changed
 - Calling .project() on a Point/Vertex will add the new geometry instead of replacing it.